requirements.txt
faster-whisper
openwakeword
piper-tts
ollama
chromadb
langchain
langchain-community
duckduckgo-search
playwright
google-api-python-client
google-auth-oauthlib
O365
psutil
fastapi
uvicorn
python-dotenv
torch
torchaudio
transformers
numpy
sounddevice
soundfile
pyaudio
webrtcvad-wheels
librosa
pyttsx3
pynput
pywin32; platform_system=="Windows"
